requirements.txt
tweepy==4.12.1
python-dotenv==0.21.0
openai==0.25.0